created get_native_namespace
